# Desktop/彼雅特github/bityacht-exchange-api-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  bankbranchs.Branch:
    properties:
      chinese:
        description: 分行名稱（中文）
        type: string
      code:
        description: 分行代號
        type: string
      english:
        description: 分行名稱（英文）
        type: string
    type: object
...
  banks.UpsertAccountRequest:
    properties:
      account:
        description: 帳號
        type: string
      banksCode:
        description: 銀行代號
        type: string
      branchsCode:
        description: 分行代號
        type: string
      coverImage:
        description: 封面照片
        type: string
      name:
        description: 帳戶名稱
        type: string
    type: object
...
  commissions.GetCommissionsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/userscommissions.Commission'
        type: array
      notWithdrew:
        description: 尚未提領
        type: number
      page:
        description: 'Default: 1'
        type: integer
      pageSize:
        description: 'Default: 25'
        type: integer
      totalReward:
        description: 累計獎勵
        type: number
    type: object
...
  errpkg.Code:
    enum:
    - 4000
    - 4001
    - 4002
    - 4003
    - 4004
    - 4005
    - 4006
    - 4007
    - 4008
    - 4009
    - 4010
    - 4011
    - 4012
    - 4013
    - 4014
    - 4015
    - 4016
    - 4017
    - 4018
    - 4019
    - 4020
    - 4021
    - 4022
    - 4023
    - 4024
    - 4025
    - 4026
    - 4027
    - 4028
    - 4029
    - 4030
    - 4031
    - 4032
    - 4033
    - 4034
    - 4035
    - 4036
    - 4037
    - 4038
    - 4039
    - 4040
    - 4041
    - 4042
    - 4043
    - 4044
    - 5000
    - 5001
    - 5002
    - 5003
    - 5004
    - 5005
    - 5006
    - 5007
    - 5008
    - 5009
    - 5010
    - 5011
    - 5012
    - 5013
    - 5014
    - 5015
    - 5016
    - 5017
    - 5018
    - 5019
    - 5020
    - 5021
    - 5022
    - 5023
    - 5024
    - 5025
    - 5026
    - 5027
    - 5028
    - 5029
    - 5030
    - 5031
    - 5032
    - 5033
    - 5034
    - 5035
    - 5036
    - 5037
    - 5038
    - 5039
    - 5040
    - 5041
    - 9993
    - 9994
    - 9995
    - 9996
    - 9997
    - 9998
    - 9999
    type: integer
    x-enum-comments:
      CodeBadErrorType: Use general error instead of *Error
    x-enum-varnames:
    - CodeBindBody
    - CodeUnauthorized
    - CodeJWTBadSigning
    - CodeJWTInvalid
    - CodeJWTRevoked
    - CodeBadAuthorizationToken
    - CodePermissionDenied
    - CodeAccountDuplicated
    - CodeBadParam
    - CodeRecordNotFound
    - CodeRecordNoChange
    - CodeBadBody
    - CodeAccountNotAvailable
    - CodeInsufficientFunds
    - CodeBadQuery
    - CodeFundDeleted
    - CodeTooManyRequests
    - CodeBadPasswordStrength
    - CodeOverResetPasswordLimit
    - CodeBadUploadedFileType
    - CodeFileNotFound
    - CodeEmailAlreadyVerified
    - CodeBadVerificationCode
    - CodeVerificationCodeExpired
    - CodeBadMobileBarcodeFormat
    - CodeBadInviteCode
    - CodeImageOverSize
    - CodeBadDataURIImageFormat
    - CodeBadImageData
    - CodeBadNationalID
    - CodeNationalIDDuplicated
    - CodeBadAction
    - CodeInsufficientBalance
    - CodeBadAmount
    - CodeTransactionPairNotFound
    - CodeCurrencyNotFound
    - CodeBadCalcOfTransaction
    - CodeBadTransactionPairStatus
    - CodeTemporaryForbidden
    - CodeBadCSVContent
    - CodeMobileBarcodeNotExist
    - CodeBadCryptocurrencyAddress
    - CodePhoneNumberDuplicated
    - CodeOverWithdrawalWhitelistLimit
    - CodeTooManyFiles
    - CodeSql
    - CodeJWTIssueToken
    - CodeGetClaims
    - CodeBadClaimsType
    - CodeSendEmail
    - CodeJWTBadPayload
    - CodeEncryption
    - CodeGormScan
    - CodeGormValue
    - CodeNewHTTPRequest
    - CodeDoHTTPRequest
    - CodeReadHTTPResponse
    - CodeSendSMS
    - CodeJSONMarshal
    - CodeJSONUnmarshal
    - CodeBadAmountOfFundsInTransactions
    - CodeBadRecord
    - CodeSaveUploadedFile
    - CodeRedis
    - CodeRedisBadScript
    - CodeCallBinanceAPI
    - CodeParseURL
    - CodeCallMaxAPI
    - CodeJSONDecode
    - CodeParseIP
    - CodeLookUpMMDB
    - CodeCallEZReceiptAPI
    - CodeLoginEZReceiptAPI
    - CodeBadParamEZReceiptAPI
    - CodeWriteCSV
    - CodeCallKryptoGOAPI
    - CodeExecuteTemplate
    - CodeUpdateCache
    - CodeCallCybavoAPI
    - CodeCybavoWalletNotFound
    - CodeWalletAddressNotGen
    - CodeWalletAddressAlreadySet
    - CodeGenQRCode
    - CodeAddressDeploying
    - CodeMemoryCacheError
    - CodeBadBase32String
    - CodeGenerateHOTP
    - CodeBadErrorType
    - CodeFailedToGenerateID
    - CodeScheduleJobKeyDuplicated
    - CodeBadCoding
    - CodeNotInit
    - CodeNotImplement
    - CodeUnknown
  errpkg.JsonError:
    properties:
      code:
        $ref: '#/definitions/errpkg.Code'
      data: {}
      error:
        type: string
    required:
    - code
    type: object
...
  sqlcache.BankInfo:
    properties:
      branchs:
        items:
          $ref: '#/definitions/bankbranchs.Branch'
        type: array
      chinese:
        description: 銀行名稱（中文）
        type: string
      code:
        description: 銀行代號
        type: string
      english:
        description: 銀行名稱（英文）
        type: string
    type: object
  sqlcache.BankOptionsResponse:
    properties:
      banks:
        items:
          $ref: '#/definitions/sqlcache.BankInfo'
        type: array
    type: object
...
  userscommissions.Action:
    enum:
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - ActionDeposit
    - ActionWithdraw
  userscommissions.Commission:
    properties:
      account:
        description: 帳號
        type: string
      action:
        allOf:
        - $ref: '#/definitions/userscommissions.Action'
        description: |-
          類型
          * 1: 返佣
          * 2: 提領
      amount:
        description: 數量 (USDT)
        type: number
      createdAt:
        description: 日期
        format: dateTime(YYYY/MM/DD HH:mm:SS)
        type: string
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: This is a RESTful API documentation of BitYacht Exchange Backend.
  title: BitYacht Exchange API Doc
  version: 0.1.0
paths:
...
  /user/banks/account:
    delete:
      description: 刪除銀行帳戶
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      security:
      - BearerAuth: []
      summary: 刪除銀行帳戶
      tags:
      - User-MemberCenter
    put:
      description: 更新銀行帳戶
      parameters:
      - description: Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/banks.UpsertAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      security:
      - BearerAuth: []
      summary: 更新銀行帳戶
      tags:
      - User-MemberCenter
  /user/banks/options:
    get:
      description: 取得銀行&分行選項
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlcache.BankOptionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      security:
      - BearerAuth: []
      summary: 取得銀行&分行選項
      tags:
      - User-MemberCenter
...
  /user/commissions:
    get:
      description: 取得返佣與提領紀錄
      parameters:
      - description: 'Default: 1'
        in: query
        name: page
        type: integer
      - description: 'Default: 25'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commissions.GetCommissionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      security:
      - BearerAuth: []
      summary: 取得返佣與提領紀錄
      tags:
      - User-MemberCenter
  /user/commissions/withdraw:
    post:
      description: 提領返佣
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      security:
      - BearerAuth: []
      summary: 提領返佣
      tags:
      - User-MemberCenter
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"